[#941] refactor: remove needed parts of typing module to abg
