+++ automation/action-binding-generator/src/main/kotlin/io/github/typesafegithub/workflows/actionbindinggenerator/TypesYamlReading.kt
@@ -1,0 +1,21 @@
+package io.github.typesafegithub.workflows.actionbindinggenerator
+
+private fun ActionType.toTyping(fieldName: String): Typing =
+    when (this.type) {
+        ActionTypeEnum.String -> StringTyping
+        ActionTypeEnum.Boolean -> BooleanTyping
+        ActionTypeEnum.Integer -> {
+            if (this.namedValues.isEmpty()) {
+                IntegerTyping
+            } else {
+                IntegerWithSpecialValueTyping(typeName = name, specialValues = namedValues)
+            }
+        }
+        ActionTypeEnum.Float -> FloatTyping
+        ActionTypeEnum.List ->
+            ListOfTypings(
+                delimiter = separator,
+                typing = listItem?.toTyping(fieldName) ?: error("Lists should have list-item set!"),
+            )
+        ActionTypeEnum.Enum -> EnumTyping(typeName = name, items = allowedValues)
+    }
