+++ automation/action-binding-generator/src/test/kotlin/io/github/typesafegithub/workflows/actionbindinggenerator/ClassNamingTest.kt
@@ -1,6 +1,5 @@
 package io.github.typesafegithub.workflows.actionbindinggenerator
 
-import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
