+++ automation/typings/build.gradle.kts
@@ -1,8 +1,7 @@
 plugins {
     buildsrc.convention.`kotlin-jvm`
-    kotlin("plugin.serialization")
 }
 
 dependencies {
-    implementation("com.charleskorn.kaml:kaml:0.55.0")
+    implementation(projects.automation.actionBindingGenerator)
 }
