+++ automation/typings/src/main/kotlin/io/github/typesafegithub/workflows/actionsmetadata/model/ActionBindingRequest.kt
@@ -1,5 +1,8 @@
 package io.github.typesafegithub.workflows.actionsmetadata.model
 
+import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
+import io.github.typesafegithub.workflows.actionbindinggenerator.Typing
+
 sealed interface TypingsSource {
     data class CodeGenerator(val inputTypings: Map<String, Typing> = emptyMap()) : TypingsSource
     object ActionTypes : TypingsSource
