+++ automation/typings/src/test/kotlin/io/github/typesafegithub/workflows/actionsmetadata/AddDeprecationInfoTest.kt
@@ -1,7 +1,7 @@
 package io.github.typesafegithub.workflows.actionsmetadata
 
+import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
 import io.github.typesafegithub.workflows.actionsmetadata.model.ActionBindingRequest
-import io.github.typesafegithub.workflows.actionsmetadata.model.ActionCoords
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
